avoid warnings, include map.h for prototypes
git-svn-id: http://svn.osgeo.org/mapserver/trunk@3643 7532c77e-422f-0410-93f4-f0b67bdd69e2

+++ cgiutil.c
@@ -3,6 +3,7 @@
 #include <string.h>
 #include <ctype.h>
 #include "cgiutil.h"
+#include "map.h"
 
 
 #define LF 10
@@ -29,7 +30,7 @@ static char *readPostBody( cgiRequestObj *request )
             exit( 1 );
         }
 
-        if (fread(data, 1, data_max, stdin) < data_max ) {
+        if ( (int) fread(data, 1, data_max, stdin) < data_max ) {
             printf("Content-type: text/html%c%c",10,10);
             printf("POST body is short\n");
             exit(1);
